Efl.Ui.Progressbar: fix val to be val when legacy format_function_set() is called
In Pb, the legacy units_format_set's user callback uses value to
be 100*value (0.0 to 100.0) and legacy format_function_set uses
value of range "0.0 to 1.0". This was broken after my patch.
Lets keep this behaviour in legacy APIs.

In case of new EO APIs, the value will be always from 0.0 to 1.0
in both format_string() and format_cb callbacks.