Revert "ecore_wl2: Remove just the flush from the idle handler"
This reverts commit e67aa661a0077f4069b67eefcb016424a630ab55.

I accidentally flagged this as a fix when it should've been a ref,
the bug (T6250) this was related to was actually fixed by the
previous 1.21 commit (a revert, returning the code to its 1.20 state).
The bug didn't exist on stable branch.

The patch was actually an optimization that requires all of my
1.21 wayland work. Ironically, backporting it to stable will
introduce the bug in T6250 there.

@ref T6250