Ui text: add ability to set item providers
Added API to set an item provider factory object.
This is similar to the previous item providers that used callbacks.

You instantiate a factory object and set it on the Ui.Text object.
Each factory implements the "create" method from
Efl.Canvas.Text.Factory.

This also includes 2 public factories (Image and Emoticon), and one
internal for Ui.Text as a fallback. The fallback provider can be
disabled via API.

See the added "Ui text factory" test in elementary_test for an
example of usage.

@feature