eldbus: "fix" make check
As Cedric told me, eldbus_suite now fails on this line, where somehow
the properties array contains 2 elements, instead of 0 as expected. It
seems that the change is not related to EFL, but a new package on our
systems.

With d-feet (a dbus inspection tool), I can see two properties under:
 org.freedesktop.DBus
   /org/freedesktop/DBus or /
     org.freedesktop.DBus
       Properties
         Features
         Interfaces

Has anyone a better clue what's happening?