edje: Use function pointer for signal cb
This introduces an obscure connection wrapper inspired from the event
callback mechanism in C++. The aim is NOT to use it as is in C, instead
C has the unchanged signal API that wraps around this. But this should
nicely support signal callbacks in all languages "out of the box"
provided they can easily keep a copy of the connection descriptor and
function pointers work well.