Canvas image: fix endless call to 'load_error_get'
Fixes a bug introduce with c8c4572d7005d19cf50f8c53bb7408606cf3ee40.
My guess is that the true intention was to query the Efl.File mixin
first, and the check the specific load_error.

Basically a call to 'efl_file_load_error_get' on an
Efl.Canvas.Image triggered this.